Fix sign conversion compiler warnings
Former-commit-id: 63c91714d2b3ab0f6cb50ae59cd861d3f370301c

+++ src/game-boy-emulator/constants.h
@@ -22,8 +22,8 @@ constexpr size_t SCREEN_RES_WIDTH = 160;
 constexpr size_t SCREEN_RES_HEIGHT = 144;
 
 // Bytes representing a 8x8 tile
-constexpr int BYTES_PER_TILE = 16;
-constexpr int PIXELS_PER_TILE = 8;
+constexpr size_t BYTES_PER_TILE = 16;
+constexpr size_t PIXELS_PER_TILE = 8;
 
 constexpr size_t BACKGROUND_SIZE_PIXELS = 256;
